requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.0.1
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.17.3
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
oauthlib==3.2.2
packaging==22.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
yarl==1.8.2
